requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.4
httpx==0.27.0
python-dateutil==2.9.0.post0
pandas==2.2.2
google-generativeai==0.8.3
python-multipart==0.0.9
